Re-added old Registration Screen
From previous MAD year, I have re-added the UserRegistry and Home pages. When the user creates an account, it sends them to the Home page to make sure that the screen routes and async storage works well with API, this user should re-appear within MySQL Database if the HTTP code 201 is returned via the API Terminal window.